CS-5001-AI-Augmented-SE/Week_2/rag_demo/requirements.txt
ollama
faiss-cpu
numpy
langchain